Bug #49609: Update byte compiled python files after upgrade from 4.3
Due to the same timestamp in 4.3 and 4.4 in the changelog for this
package, the byte compiled python files (*.pyc) were not updated by
pysupport. That resulted in multiple tracebacks due to python still
using the old byte compiled files, because the timestamp was the same
for python files from this package in 4.4 and 4.3.