Bug #55252: Add feedback link to every heading
Univention Sphinx extension adds a link to the [Univention feedback
form](https://www.univention.com/feedback/) to every heading. It allows
readers to easier provide distinct feedback. The feedback already
contains information about the location the feedback refers to.

The feedback link is only applied to headings in the HTML build. The PDF
build artifact is not touched.