Issue hackathon#62: using docker-registry.knut.univention.de/knut/pre-commit-kelvin as CI_REGISTRY_IMAGE
Changed the base from .kaniko to .kaniko_knut in build_pre_commit to allow the change of CI_REGISTRY_IMAGE.
Setting CI_REGISTRY_IMAGE: $CI_REGISTRY_IMAGE/pre-commit did not work because it yields an UNAUTHORIZED error because of the wrong $CI_REGISTRY.

If CI_REGISTRY_IMAGE is set as in build_pre_commit, it has to be re-set per job. This is why it has to be set in build_docker_image.