Bug #55986: fix is_x checks for User
We discovered while working on Bug #55986 that the `is_x` checks
(`is_student`, `is_administrator`) do not work correctly for the User in
the `ucs-school-lib` because the code is based on objects retrieved
directly from UDM instead of the UDM REST API. Consequently, it throws
an error when it tries to access `oldattr`.

This commit updates the code to make sure that when those functions are used,
they won't throw an error.