perf(ucs-school-lib): Do not fetch non-school group objects from UDM
During the modification of a user, we iterate through all of his
UDM groups and check if he should be removed from them.

This is limited to groups which are school groups (school classes and
work groups) and hence we do not need to fetch the UDM group object,
like it was done before this performance improvement.

Bug #58475
Issue univention/dev/education/ucsschool-kelvin-rest-api#158